work with layouts

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -101,6 +101,11 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        $category = Category::find($id);
+        if($category->posts->count()) {
+            return redirect()->route('categories.index')->with('error', 'К посту привязаны категории');
+        }
+
         $category = Category::destroy($id);
         return redirect()->route('categories.index')->with('success', 'Категория удалена');
     }
